fix(edit): do not treat the texture LRU budget as live-allocation failure
Peak in-flight pipeline textures can exceed the 256 MiB LRU target while
device memory is still available. Keep valid Develop leased and fail only
when the backend cannot create a texture.
